+++ src/ERM19264_graphics.h
@@ -1,7 +1,6 @@
 /*
 * Project Name: ERM19264_UC1609
 * File: ERM19264_graphics.h
-* Description: ERM19264 LCD driven by UC1609C controller header file for thegraphics functions based on Adafruit graphics library
 * Author: Gavin Lyons.
 * URL: https://github.com/gavinlyonsrepo/ERM19264_UC1609
 */
@@ -18,9 +17,33 @@
 
 #define swap(a, b) { int16_t t = a; a = b; b = t; }
 
-#define ERM19264_ASCII_OFFSET 0x00
-#define ERM19264_ASCII_OFFSET_SP 0x20 // Starts at Space
-#define ERM19264_ASCII_OFFSET_NUM 0x30 // Starts at number 0
+typedef enum 
+{
+    UC1609Font_Default = 1,
+    UC1609Font_Thick = 2,
+    UC1609Font_Seven_Seg = 3,
+    UC1609Font_Wide = 4,
+    UC1609Font_Bignum = 5,
+    UC1609Font_Mednum = 6,
+}LCD_FONT_TYPE_e;
+
+typedef enum 
+{
+	FONT_W_5 = 5, FONT_W_7 = 7, FONT_W_4 = 4, FONT_W_8 = 8,FONT_W_16= 16
+}LCD_Font_width_e; // width of the font in bytes cols.
+
+typedef enum 
+{
+	FONT_O_EXTEND = 0x00, //   extends ASCII
+	FONT_O_SP = 0x20,  // Starts at Space
+	FONT_O_NUM = 0x30,  // Starts at number '0'
+}LCD_Font_offset_e; // font offset in the ASCII table
+
+typedef enum 
+{
+	FONT_H_8 = 8, FONT_H_16 = 16, FONT_H_32 = 32
+}LCD_Font_height_e; // width of the font in bits
+
 
 class ERM19264_graphics : public Print {
 
@@ -53,8 +76,6 @@ class ERM19264_graphics : public Print {
 	void	fillRoundRect(int16_t x0, int16_t y0, int16_t w, int16_t h,
 			int16_t radius, uint8_t color);
 	void	drawBitmap(int16_t x, int16_t y, const uint8_t *bitmap,
-			int16_t w, int16_t h, uint8_t color);
-	void	drawBitmap(int16_t x, int16_t y, const uint8_t *bitmap,
 			int16_t w, int16_t h, uint8_t color, uint8_t bg);
 			
 	void	drawChar(int16_t x, int16_t y, unsigned char c, uint8_t color,
@@ -64,8 +85,9 @@ class ERM19264_graphics : public Print {
 	void	setTextColor(uint8_t c, uint8_t bg);
 	void	setTextSize(uint8_t s);
 	void	setTextWrap(boolean w);
-
-	void setFontNum(uint8_t FontNumber);
+	void	setDrawBitmapAddr(boolean mode);
+	
+	void setFontNum(LCD_FONT_TYPE_e FontNumber);
 	void drawCharNumFont(uint8_t x, uint8_t y, uint8_t c, uint8_t color , uint8_t bg);
 	void drawTextNumFont(uint8_t x, uint8_t y, char *pText, uint8_t color, uint8_t bg);
 
@@ -77,26 +99,30 @@ class ERM19264_graphics : public Print {
 
 	int16_t height(void) const;
 	int16_t width(void) const;
+    
+	LCD_FONT_TYPE_e FontNum;
 
  protected:
 	const int16_t WIDTH;
-	const int16_t	HEIGHT;  
+	const int16_t HEIGHT;  
 
 	int16_t _width;
 	int16_t _height;
 	int16_t cursor_x;
 	int16_t cursor_y;
 	
-	uint16_t textcolor;
-	uint16_t textbgcolor;
+	uint8_t textcolor;
+	uint8_t textbgcolor;
 	
 	boolean wrap; // If set, 'wrap' text at right edge of display
-	
+	boolean drawBitmapAddr; // True vertical , false = horizontal
+		
 	uint8_t   textsize;
-	uint8_t _FontNumber = 1;
-	uint8_t _CurrentFontWidth = 5;
-	uint8_t _CurrentFontoffset = 0;
-	uint8_t _CurrentFontheight = 8;
+	uint8_t _FontNumber = UC1609Font_Default;
+	uint8_t _CurrentFontWidth = FONT_W_5;
+	uint8_t _CurrentFontoffset = FONT_O_EXTEND ;
+	uint8_t _CurrentFontheight = FONT_H_8;
+	
 };
 
 #endif