+++ extras/doc/CHANGELOG.md
@@ -16,3 +16,12 @@
 
 * Version 1.3.0 June 2021
 	* Added five fonts.
+
+* Version 1.4.0 Feb 2022
+	* drawBitmap() function is now set up for both horizontal and vertical addressed bitmap data,
+	as per pull request  #2 from Myaflick on github.  Vertical is default. The addressing mode is changed by setDrawBitmapAddr(), new function.  
+	* Enum added for font name labels instead of raw numbers, This will cause font compiler warnings for sketch's written on versions before 1.4.0 , The sketch's will still compile and work,
+	Simply replace font numbers  with the relevant enum text labels to get rid of warnings. 
+	* A new function added for initialising a multibuffer struct , LCDinitBufferStruct(),
+	backward compatible with old manual method.  
+